Improvements (#54)
* Update queries.

* Wrap API error with response class and data result with Either class.

* Enable debugging.

* Create API response wrapper class.

* Fix episode images not loading and cleanup season and episode queries

* Set hardwareAccelerated to true

* Wrap API calls with ApiResponse class

* Update queries.

* Cleanup: Update repositories, cleanup mappers and mockdata, remove unused functions.

* Add generic exception.

* Update states.

* Update TopBarView and add animation

* Minor cleanup

* Minor cleanup

* Update detailView. Fix stretch down gesture and add toolbar title animation.

* Minor cleanup.

* Delete test module.

* Switch to runTest and cleanup tests.

* Spilt up trakt repository and to profile and show repository.

* Use correct repository after splitting it up.

* Update query

* Update error component.

* Update states.

* Remove enableFeaturePreview

* Update gradle build and test commands.

* Update gradle wrapper version.

* Disable hierarchicalStructureSupport

* Build src migration: Herding Elephants. (#53)

* Get rid of buildSrc and pre-compiled scripts

* create gradle plugins.

* Switch to plugin scripts and cleanup dependencies.

* Update integers to long

* Update import after upgrading sqlDelight

* Create kmm scripts

* Provide coroutineContext

* Provide coroutineContext

* Use kmm plugins & update dependencies.

* Delete unused class

* Cleanup: Move common block to script

* Cleanup: update gradle scripts.

* Minor cleanup: add manifestPlaceholders, add suppress annotation.

* Add app flavor extension.

* Simplify build job

* Remove unused blank sourceSets

* Remove unused dependency

* Update lib

* Add flow preview to opt-in.

* Update dependencies.

* Add artwork

* Update readme.

* Propagate error downstream

* Catch exception.

* Remove network exception and rename class.

* Add helper function and optIn annotations.

* Minor cleanup

* Add optIn annotation.

* Update gradle wrapper version.

* Combine show detail requests into one call and update UI states.

* Fix failing test.

* Move modules from domain to data module.

* Refactor and rename show module to category module:

- Move show related implementation classes to tmdb and trakt module.

* Remove database dependency.

* update dependencies.

* Add domain plugin.

* Split up module to data and domain.

- Statemachine implementation has been moved to domain module. This might be renamed to presentation later.

* Delete shows module.

* Update imports.

* Update module dependencies.

* Use expect actual pattern to add koin injection support.

* Use expect actual pattern to add koin injection support.

* Update imports and add data and domain modules.

* Rename package name

* Update setup instructions: Use JDK 17 instead of 11

* Minor cleanup: Rename class and update imports.

* Update modules.

* Load individual items independently preventing blocking the UI

* Delete unused classes.

* Rename episode table id column

* Rename function

* Add functions to save and get season id.

* Some minor cleanup.

* Update JDK to 17 and cancel jobs is there's a new one

* Change EpisodeArtWork return type to flow. This fixes episode images not being updated.

* Add coroutine parameter.

* Use datastore to save and retrieve the seasonId.

* Update dependencies.

* Minor cleanup: Change class to interface

* Minor cleanup: Get rid of mutableStateFlow and move code blocks to functions.

* Add error test case and fix failing test.

* Set minimum height to episode item

+++ android/features/following/src/main/kotlin/com/thomaskioko/tvmaniac/following/FollowingListScreen.kt
@@ -24,17 +24,17 @@ import com.thomaskioko.tvmaniac.compose.components.AsyncImageComposable
 import com.thomaskioko.tvmaniac.compose.components.EmptyContentView
 import com.thomaskioko.tvmaniac.compose.components.ErrorUi
 import com.thomaskioko.tvmaniac.compose.components.FullScreenLoading
-import com.thomaskioko.tvmaniac.domain.following.api.ErrorLoadingShows
-import com.thomaskioko.tvmaniac.domain.following.api.FollowedShow
-import com.thomaskioko.tvmaniac.domain.following.api.ReloadFollowedShows
-import com.thomaskioko.tvmaniac.domain.following.api.FollowingContent
-import com.thomaskioko.tvmaniac.domain.following.api.LoadingShows
+import com.thomaskioko.tvmaniac.domain.following.ErrorLoadingShows
+import com.thomaskioko.tvmaniac.domain.following.FollowedShow
+import com.thomaskioko.tvmaniac.domain.following.FollowingContent
+import com.thomaskioko.tvmaniac.domain.following.LoadingShows
+import com.thomaskioko.tvmaniac.domain.following.ReloadFollowedShows
 import com.thomaskioko.tvmaniac.resources.R
 
 @Composable
 fun FollowingScreen(
     viewModel: FollowingViewModel,
-    openShowDetails: (showId: Int) -> Unit,
+    openShowDetails: (showId: Long) -> Unit,
 ) {
 
     val followedState by viewModel.state.collectAsStateWithLifecycle()
@@ -77,7 +77,7 @@ fun FollowingScreen(
 private fun FollowingGridContent(
     list: List<FollowedShow>,
     paddingValues: PaddingValues,
-    onItemClicked: (Int) -> Unit,
+    onItemClicked: (Long) -> Unit,
 ) {
     val listState = rememberLazyListState()
